requirements.txt
pandas
numpy
torch
scikit-learn
python-dotenv
requests
openpyxl
tqdm
joblib
transformers
Flask
